requirements.txt
Django==5.0.0
Pillow==10.1.0
rembg==2.0.57
python-magic==0.4.27
